requirements.txt
Django==2.0.6
djangorestframework==3.8.2
pylint==1.9.2
django-nested-admin==3.0.21
social-auth-app-django==2.1.0
drf-nested-routers==0.90.2
